# Conflicts:
#	src/ShapeCrawler/ShapeCollection/SlideShapes.cs

+++ src/ShapeCrawler/Tables/BottomBorder.cs
@@ -31,7 +31,7 @@ internal class BottomBorder : IBorder
             this.aTableCellProperties.BottomBorderLineProperties.AppendChild(aSolidFill);
         }
         
-        var emus = new Points(points).AsEmus();
+        var emus = new Points((decimal)points).AsEmus();
         this.aTableCellProperties.BottomBorderLineProperties!.Width = new Int32Value((int)emus);
     }
