+++ src/ShapeCrawler/Tables/RightBorder.cs
@@ -31,7 +31,7 @@ internal class RightBorder : IBorder
             this.aTableCellProperties.RightBorderLineProperties.AppendChild(aSolidFill);
         }
         
-        var emus = new Points(points).AsEmus();
+        var emus = new Points((decimal)points).AsEmus();
         this.aTableCellProperties.RightBorderLineProperties!.Width = new Int32Value((int)emus);
     }
